Add Update Games List button to Gaming Tab
Add a new 'Update Games List' button to the Gaming Tab that launches the
Extractor dialog configured to use the --update-games flag. This allows
users to easily update the gamedata.txt file from TrackIR installers
without extracting firmware files.

Key changes:
- UI: Added horizontal layout with PrereqGroupBox and UpdateGamesButton
- Extractor: New UpdateGamesExtractor and UpdateGamesExtractThread classes
- Memory: Fixed memory leak by converting static pointer to member variable
- Cleanup: Added proper initialization and destructor cleanup

The button is placed to the right of the Prerequisites group box, which is
resized to half width to accommodate the new button.